Fix vector variable diagnostic display
There were two problems. The first was simply that I forgot to pass on
the variable type to the LoopDisplay component, so it was always trying
to display a scalar variable; even for wind.

The second was a little more odd. Svelte was setting a data attribute,
not a data property on the chart-2dhistogram component. I had to
override the data property on Chart2DHistogram and then Svelte started
doing the right thing.

Most likely Svelte is checking to see if the property is defined on the
class — hasOwnProperty style — and if not, assuming it's an attribute. I
think Preact was probably just testing for the existence of the property
anywhere in the class hierarchy. Not sure which is better, since
Svelte's approach might prevent you from accidentally making weird
assignments to built-in Object properties, but also undermines some of
the value of inheritance. Ahhh, JavaScript…