Create a new component for managing charts and axes
Instead of having a component that all charts inherit from to manage
resizing, it makes more sense to me to have a container that uses slots
for the chart itself and the axis titles. Composition like this will
make things a little more testable, and perhaps slim down some of the
logic in the charts themselves.