Reduce ChartElement to just some default props
The base ChartElement no longer provides any template markup or resize
handling. It exposes a height and width attribute (and property) that
trigger a re-render. The width and height can be set by putting the
chart inside a ChartContainer.

This should make it easier to use the base ChartElement with the map as
well as the SVG charts.