Remove undefined references to domain/range on heatmap
The 2D histogram was trying to read the domain and range for the chart
off of some properties that don't exist, and then falling back to
calculating it from data. For now I've removed the properties and we
just set it from the data, but we will want to be able to specify a
domain and range for the chart just so that we can create multiple
charts on the same axes.