Update pytest to the latest
pytest 5.4 had a dependency on the "py" library. This was triggering our
vulnerability scanners. The latest version of pytest has dropped the
dependency on "py" so will fix the issue. This is a dev dependency so
there is some debate on if this is really necessary. However, it seems
important to get the container deployed before the demo and this is the
easiest path forward.