Add a scroll-container utility
I had removed the utilities module because trying to use USWDS Sass
tokens was causing exceptions. I've removed the import of uswds-core so
that we can include this once again. It's confusing to have it in the
project and not included, I just spent an hour trying to figure out why
the scroll-container wasn't scrolling; it was because the rule wasn't
included in our CSS output.

I removed all of the unused rules from utilities and replaced them with
a scroll-container definition that makes it easy to create our
scrollable areas.