Change overflow to clip for html and body
I recently learned the difference between `overflow: hidden` and `clip`.
`hidden` is still scrollable programmatically, so it is still a scroll
container, whereas clip is not at all scrollable. I don't think this
matters for us in this case, but `clip` is the intention: the document
root and body should not be scroll containers in this app.