Add unit tests to the S3 code
We're using moto in server mode as a local S3 service. This is similar
to how the s3fs library runs its tests. I ended up needing to modify the
code that opens files in S3 to allow us to inject a non-AWS endpoint for
S3 and to specify the path used for the local cache.

Making sure that fsspec.open() was being called correctly in the tests
was more involved than I would have liked. We had to make sure that AWS
was being properly mocked out and that the local cache was being cleared
in between tests.