Add a skeleton for the color-ramp component
The color ramp is another web component. It takes a title as child
content and will render the color ramp in SVG. I'm not yet sure how to
define the size of the color ramp so that you have good control over the
height of the ramp itself and the font-size of the tick labels.