Refactor variables from render() into properties
I think the way I want the color ramp to work is to give it a for
attribute that's an IDREF to the chart that it renders the scale for.
Which means the chart needs to expose its scale as a property for the
color ramp to read. Which means that all of the dependencies that go
into generating the fill scale for the 2D histogram have to be available
as properties as well.

(Maybe they don't *have* to be properties, but it makes things a little
easier to reason about, I suppose.)