Shrink the text in the color ramp
Use a utility class to make the text in the color ramp a little smaller.
This should, however, probably be a global style. All axis labels like
this should be exposed as `::part()` on the web components for easier
styling.

This font size doesn't match the font size of the axis labels on the
charts themselves because D3 "helpfully" sets font sizes and styles as
attributes on the SVG elements themselves. I'll probably have to fight
this.