Listen for chart-datachanged event to update color ramp
The map doesn't get recreated on variable changes, so we have to have a
way to signal to the color ramp that it should update. Part of me
wonders if this should be a mutation observer so that we don't have to
have each chart component fire this event when its data changes. Or
perhaps finding a way to effectively mix in a data property to all of
our charts so that the event handling exists in only one place.