Add a test for fetching just vector magnitude
We want to switch the distribution plots for vectors to display just the
u and v components of the vectors, rather than transforming them to
polar coordinates. But we will still want to plot the magnitude on the
map, so we will need a separate endpoint for that, I think.