Separate map data from distribution data
In order to support displaying wind as a vector (u, v) instead of in
polar coordinates and keep the map displaying the magnitude, we have to
separate the data for the distributions vs the map data. When the
variable is a vector, we make a separate HTTP request for
/api/diag/<variable>/magnitude/ to get the map data.

This also means that our brushing has to be based on the station ID for
the observation, because we can't map a selection in (u, v) space to the
magnitude of the vector. Instead, we have to find all of the
observations that fall within the selection, and then exclude all
observations from the map that aren't in that list.