Fix the 2D histogram
Update the 2D histogram chart to look for `u` and `v` instead of
`direction` and `magnitude`. This is not a good solution, these charts
should be more general than this, but I think I'm going to rewrite this
as a scatterplot anyway, so this will be good enough to ensure we have a
demo for January.