Handle histogram selections separately from d3
Store the selection as a private member of the ChartHistogram element,
rather than as the datum on a d3 selection. For some reason, clearing
the linked selection wasn't working when the charts were clicked — only
the clicked chart actually cleared the selection rectangle element.
Changing selection management to be the same as we do for the map has
fixed this issue.