Set the empty range value to [0, 0]
When we clear the selection by setting the range store to null, none of
the other web components' selections are updated. So we have to set it
to an actual value or the other histogram doesn't clear its selection.
We use a value of [0, 0], but we check for any selection where the lower
bound equals the upper bound to clear the selections. We'll need to do
the same for regions on the map and the 2D histogram.