Increase the contrast of the annotations against axes
Trying some quick hacks to increase the contrast of the annotation
labels against the background of the rest of the chart. They are now
bold, and I've added a white text shadow to knock out some of whatever
is behind them. This is not quite legible enough for my taste, so I may
need to do some more work on these annotation utilities, but this might
be good enough for now.