Add stubs for the data pipeline
To think through the process of converting the incoming NetCDF files to
a Zarr, I'm stubbing out the process at a high level for the Lambda
handler function. The Lambda handler just deals with translating the
arguments to the handler to something we can pass on to other functions,
per the best practices in the Lambda docs so that we can more easily
test the logic of the pipeline outside of a Lambda environment.