Refactor netcdf converter to avoid AWS
Instead of having diag.to_zarr create the Zarr array in S3, let's return
the Dataset and allow the lambda_handler to deal with uploading it to
the Zarr (or we may want to refactor that logic as well to more easily
test our group definition). This way the diag module can be tested
without needing anything for S3.

The diag module is going to handle conversion of some fields, such as
the Longitude coordinate which needs to be converted from the range [0,
360] to [-180, 180].