Ignore requirements.txt in this service
We may end up with a requirements.txt file in this service because we
generate one in the Docker image for Lambda. Since we don't actually use
this for managing requirements, we want to make sure we don't commit it
by accident if it ends up in the working directory somehow.