Add config for coverage
This will get us a more accurate report of our code coverage. (Note that
the coverage measurements went down.) We stop measuring coverage of
tests, because that's silly, and we measure coverage of all modules in
src/, whether they're executed or not. So now coverage is reporting
aws.py as having 0% coverage.