Return a list of variables parsed from the filename
To help deal with vector variables, I've changed the parse_diag_filename
function to return a list of variables it finds in the filename. Right
now it's hard-coded to return ["u", "v"] when it finds the variable "uv"
because that's the only vector we currently have. It seems reasonable to
me to construe that as a list of two variables, rather than as a single
variable, considering that's how its stored in the NetCDF file.