Fix some type errors
mypy was upset because the timestamp match might have been None,
although execution couldn't get that far because if the first regex
matched, the second regex had to match, being a subset of the first.
Nonetheless, I combined the two regexes into one and now mypy is happy
because I test for None first before extracting the groups.