Store variables in separate groups
It turns out that variables can have differing numbers of observations
even if they are from the same model run, so we can't store all of the
variables in a single group. So I had to refactor the data format to
store the obs, forecast, and difference for each variable in a separate
group, rather than storing all of the observations together, and all of
the forecasts together, etc.