Fix the open S3 diag test
I modified the test and fixtures to read from an actual S3 bucket. I
don't think we want to use moto to mock AWS because it involves
monkeypatching the boto3 client that s3fs uses. Plus, I think it's good
to have some actual integration tests with S3.

There are three challenges. One, it's slow to write the Zarr to S3 every
time. Two, we may not be able to access S3 from CI. Three, we need to
clean up the bucket in S3.