Remove parameters from open_diagnostic s3 test
I don't think there's much value in testing each variable when we read
from S3. We just want to make sure that given the correct environment,
the application is able to read data out of a Zarr in S3. Since these
tests are slow, it's better to test the minimum necessary.