Implement loops function for the list loops API
I added the /diag/<variable>/<init_time>/ route to handle fetching the
available loops. This, coincidentally, took care of the bug with the
endpoint that lacked a trailing slash. (Maybe I could've just fixed that
without breaking anything…). I had to update the URL for the diag data
otherwise it would have conflicted with the list loops endpoint, but the
get diag tests are all failing right now because that hasn't actually
been implemented.