Update the UI to retrieve init times and fetch loops
Updated the Svelte components to load the init times for each variable.
I decided to replace the model selection (which currently does nothing)
with the variable selection, because variable selection (currently) has
to happen before we know what init times are available. I think this is
backwards and probably needs to change in a future refactor.

I updated the app to fetch each loop separately, since we can't combine
loops reliably (because they may have a different number of
observations).