Apply filters to wind data
Refactor the filtering to a function that can be called for both scalar
and vector variables. I had to make an adjustment to how we index the
data when building out the observations because when there’s only a
single observation in a vector, len() returns 2 (the number of
components) instead of 1 (the number of observations).