Reverse the latitude values in tests
The frontend constructs the bounding box in such a way that the latitude
coordinates are sorted in descending order. We were assuming in the
filtering code that ranges were sorted ascending, so nothing was being
returned. Instead of assuming, we should just sort the incoming ranges
to ensure that our filters work.