Modify parameter format to support vector values
Instead of specifying ranges with a ",", we use "::" (not "-" because we
need to support negative values). The comma is now used to separate
components of a vector and we just build our filters assuming that it is
a range of vector components. Thanks to the magic of numpy, filters work
even if we're just selecting a single value.