Fix the wind mapRadius
Add a helper to calculate the magnitude of the wind vectors for
displaying on the wind map. We lose the directionality of the
observation - forecast with this because we don't have access to the
original observation and forecast values in the client.

For now I think this is fine, because I suspect, based on some images
that Terra shared of how they visualize observation sites, we may not
display data like this on the map, and instead we might show just the
sites and allow users to select them for more detail.