Separate the app and data caches
WE can employ two simple caching strategies based on the type of data we
need. For data resources, we can just cache them indefinitely. The
output of a model run shouldn't change, so once we have a copy, we can
just use that. Everything else — the app assets — we can always respond
with the cached version, but always update the cache in the background.
This way, when we push an update, users still get fast loads, and only
have to wait one page load to get the latest version of the app.