Consolidate duplicate fetches
It’s convenient to let each chart exist in isolation and just fetch its
own data, even if two charts use the same data. But it's more efficient
to ensure that we only make one actual HTTP request and just share the
response data with each requester. I've added a Map to the service
worker to track in-flight fetches and clone the response instead of
opening a new fetch request.