Expect "Object created" events from EventBridge
We're not getting the kinds of events we expected from the S3 events.
Ian thinks this may have to do with the fact that VLab had to use
EventBridge to hook up our lambda function. The format we're expecting
now is the format we logged from the handler when we didn't get what we
expected, so hopefully this works.