Add linters and formatters for CSS/JS
Copy over the package.json and linter configs from the UI service.

We're switching over to alphabetical ordering of CSS declarations
because it's just easier to maintain, and some consistent order is
better than none. I ran the CSS files through the formatter so that they
will pass the linting checks.