Fix a race condition in ChartContainer loading
When we're not bundling all of our Web Components there's a race
condition caused by upgrading chart-container before its children are
upgraded that breaks the width/height setters on all the chart elements.
So we had to add a check for custom elements (by detecting a "-" in the
tag name) and wait to set their width and height until the component is
upgraded.