Create a vendored version of D3 as an ESM
I don't want to have to run a bundler in development, but D3 doesn't
provide any bundles as ESM for us to use. We also can't resolve imports
in node_modules with Flask.

To work around this, I added an NPM script that creates a bundle from D3
for us in static/js/vendor/d3.js. Run `npm run vendor` to build the
module and then you can start the dev server. This is nice because it
lets us use NPM to manage dependencies so dependabot will still notify
us when there's an update.

We'll need to add a build step for production that bundles our code so
that we benefit from tree-shaking out all of the bits of D3 we don't
use.