Bundle and minify the CSS for production
I added PostCSS to bundle and minify our CSS in production. I also added
a second Dockerfile for the nginx server (replacing the container we
used for the old UI service). This does essentially the same thing as
the old Docker container for the separate UI service: it starts with a
Node image to bundle our CSS, then copies the built assets into a
directory for nginx. I still need to add the configuration for the
remote proxy.

We'll also need to start bundling the JS here.