Add s3fs as an explicit dependency
s3fs wasn't showing up in the requirements.txt when exported from
poetry. This was causing the lambda function to fail so add it as an
explicit dependency. s3fs uses aiobotocore as one of its dependencies
and it has very restrictive versions of boto3 it can work with. I had to
add that as a dependency and give it some very permissive version rules
in order for things to work.