Prioritize network requests for HTML documents
The HTML contains markup for things like selecting the latest model
runs, so we don’t actually want to prioritize the cache for these
requests because that will result in stale options from the menus.
Not to mention the issues caused by caching the redirect responses
(although that should be fixed).

It seems safer to prioritize the latest version of the HTML from the
network and cache it for offline use rather than using the cached
version first.