Add lodash as a dev dependency
Although it is quite a large addition to our bundle lodash.get is the
most reliable tool I know for accessing properties on an object using a
path string. Although we don’t necessarily all of the path formats it
supports, this seems like the easiest way to support configuring the
value accessor on the ChartMap for rendering O-B and O-A—at least for
now.