Extract background from the diag file names
Add support for reading optional background information from the diag
file names. I decided to drop support for the file name formats where
the init time comes after the file extension because it was a little
difficult to support in a single regex. I think we’ve only seen this
format in the first set of test files we had, and I think it’s ok to
just make it a requirement that everything comes before the file
extension.