Parse model, system, etc. from path to diag file
I added model, system, domain, and frequency to the DiagMeta namedtuple,
and parse those out of the path in parse_diag_filename. The default for
any missing tokens in the path is None. I also added support for a
prefix parameter for parse_diag_filename that is ignored if present on
the filename. It’s an error if the prefix is specified and not present
or doesn’t match.