Rename all of the metadata listing functions
Rather than using plurals like `models` or `systems` the functions are
called `get_model_list` or `get_system_list`, which makes it easier for
us to generate the function name from a noun like "model" or "frequency"
without having to deal with the fact that pluralizing "frequency" is not
the same operation as pluralizing "model". Which makes it easier for us
to handle metadata processing in loops.