Change metadata handling to parse just filenames
When we download the NetCDF files from S3, we combin the path-like
tokens from the key with the filename using underscores. So I had to
update all of the parsing logic to read out of the filename, not expect
a path. I’ve successfully used the handler function to download test
data from S3, so I’m more confident now that this is working.