Update tests to expect the background in the URL
The background is now being encoded in the Zarr group path for each
array, so we need to accommodate that in both the URLs and the diag
reading code. This gets all of the old tests passing with this new path.
I also removed one test case for an unknown variable test because I
think it was just providing an invalid URL that should 500, not 404.

The tests for listing each piece of metadata based on values further up
the path are all failing, but they will be removed (at least for now)
because I think we can just splat all of the possible values into the
form and 404 if someone picks a combo that doesn’t exist.