Fix selection in the 2D histogram
The code was still using the private xScale and yScale variables to
invert the coordinates in the UI, but those aren’t being used by the
actual drawing code, so they were just 1::1 with the screen coordinates.
Those references to the private scale variables have been replaced with
the getter for the scale.

The format of the custom event’s detail has also been changed to be more
useful to the event handler. The event handler has been updated to
handle vector selections as well.