Move heading styles into utility classes
We’re eschewing global element styles for things like headings to make
it easier to construct a document outline while maintaining an
appropriate visual hierarchy in the app. Since we’re not typesetting any
prose, it will be easier to do this with classes and avoid having to
battle element styles.