Add a new fixture for test data
I think I’ll want to replace some of the existing fixtures with a
combination of this one, and a reduced use for diag_zarr. This should
make it easier to generate arbitrary test data for any of our test cases
and might make it easier to write more comprehensive and comprehensible
tests.