Switch to the nginx-unprivileged image
We won't be able to run OS updates since the image uses the nginx user.
Updated images are released weekly on Mondays so hopefully that will be
sufficient. The unprivileged image also uses 8080 as the default port so
we needed to explicitly expose port 80 to avoid having to change the
rest of our nginx and kubernetes configs.