Move the ETL modules into the API service
Move the ETL (extract, transform, load) modules from the data service to
unified_graphics.etl to consolidate all of our applications. This will
make it possible to expose command line tools in the application via the
`flask` command to load data (which will be handy for back-filling and
for development), will allow us to share SQLAlchemy models when we add a
database for metadata (#300), and allows us to eliminate duplicate test
fixtures.