Downgrade to Python 3.9 and update dependencies
Added missing dependencies needed by the ETL pipeline and modified some
of the version requirements for existing dependencies to accommodate
botocore and boto3.

**Downgraded from Python 3.11 to 3.9.** The available Lambda docker
images from Amazon only support Python 3.9, so we are downgrading the
application Python version to 3.9 as well. We could, in the future, look
into using Nox or Tox to test against multiple versions of Python to
allow us to use 3.11 for the application, but I’m not sure we get much
of a win, since we still have to target 3.9 which means we miss out on
things like the new match statement anyway.