Move the Dockerfile for Lambda to the API service
Rename the Dockerfile for the Lambda function to Dockerfile-diag-etl in
the API service. Update it to run the handler function from the
unified_graphics package instead of ugdata.

I’m pretty sure this is all we need to do to get this working. It feels
a little bad installing the whole unified_graphics package in the
container for the Lambda function, since it doesn’t need Flask or
anything like that. Maybe we can look into groups in poetry to at least
reduce the size of the dependencies.