Run isort over GitHub scripts
Because we’ve moved all of our code to the project root and are now
running checks there, we’re catching issues in non-application code,
like the import order of our GitHub scripts. I see no harm in applying
the same formatting requirements to our tools that we apply to the
application, so I went ahead and sorted the imports.