Fix queries for models and backgrounds
Instead of returning the full list of models for both model runs and
backgrounds, we can query for all the models used in an analysis for the
models, and query for all the models used as backgrounds for the
backgrounds. This gives us a better list for the UI and keeps our test
conditions the same, so the API hasn’t changed in any way.

We should probably create some materialized views for these queries and
provide models for them to make these queries easier to run.