Return init times as strings for metadata
Since we’re using the get_model_metadata response to build forms, we
might as well format the datetime objects the way we format our times.
Arguably this should be handled by the template for increased
flexibility with this function, but for now my concern is just returning
the app to its previously working state using the database to speed
things up.