Remove Flask-Migrate
Flask-Migrate makes a flask app context a requirement for running
migrations because of how it sets up alembic. So instead I removed it as
a dependency and just added alembic directly. There wasn’t much extra
work to get alembic configured to look for the
FLASK_SQLALCHEMY_DATABASE_URI variable. I did have to move the
flask_sqlalchemy.SQLAlchemy object to unified_graphics.models to ensure
that when it's imported, all of the models get registered on it so that
alembic detects them. That required updating a few imports elsewhere in
the package as well.