Allow programmitic config to override environment variables
Swap the order of configuration back to environment first and then
config passed in to the app factory. I'm not entirely convinced this is
the right way to do this, but it is necessary right now for ensuring
that we can configure the test environment by overriding any environment
variables when developers have a .env file in their working tree.
Otherwise the dev config in .env always trumps the test config an tests
fail and the dev database is polluted.