Eliminate dependency on app context in fixtures
I rewrote some of the fixtures so that they don't rely on an app context
except when it is actually necessary (such as the test client). We were
requiring an app context just to read the DIAG_ZARR config option, but
we could make the path to the DIAG_ZARR into its own fixture and then
make that a dependency for the app fixture. That way anything that just
needs to read that path can use the fixture instead of getting an app
and creating an app_context.

The hope here is that this helps eliminate the errant database
connections that are causing errors at the end of the test suite when
tearing down the test database; though so far that doesn't seem to be
the case. Nonetheless, this feels a little less convoluted.