Eliminate Flask app from diag module
Attempting to fix this database clean up problem in our tests by
removing the dependency on Flask from the diag module. Instead of using
current_app and db directly, database sessions and paths are passed in
by the callers. This shifts the dependency on the Flask app to the
routes module, which makes sense, since they define Flask-specific
functions. This makes it easier to test the diag module directly because
we don’t have to set up a Flask app context to run them, we can just
pass in the needed values when we run tests.

This does not, however, fix the database cleanup error.