Separate test db creation from test engine fixture
In an effort to follow pytest recommendations, I've separated the
fixture that creates the test database from the fixture that creates the
engine connected to that database. The test db fixture yields the URL
for the test database, which is then used by the engine and app fixtures
to define the engines for the test database.