Kill any sessions before dropping the test db
Well, going with the nuclear option here. Before we try to drop the test
database, we just kill off any existing connections. I still don't know
where this connection is coming from. The code to kill the connections
was cribbed from the SQLAlchemy-Utils package (which we aren't using,
because we need very little from it).

It's possible that this is just an issue with PostgreSQL. The code in
the SQLAlchemy-Utils package does this only for PostgreSQL databases.