Configure test DB connection through environment
Use environment variables to configure the username, password, and host
for the test database. This will allow us to configure the connection to
the test database in CI. The test config runs dotenv if it's available
just to make it easy to configure these variables in development, and
there are also some reasonable defaults for the environment variables.