Disable tracebacks in the CI test output
Generally these are not that helpful, especially if the error is a SQL
error; we get long traces into SQLAlchemy's code and it just obscures
any other information in the output. In CI, I think, all we need to know
is what failed/errored; we can run the tests locally for more details.