Test against database for ETL save tests
Add assert statements to the test_save tests to ensure that we are
creating one and only one record for each analysis when we ingest a new
file.

I refactored the tests a little to make them easier to follow (I hope)
using fixtures to encapsulate the actions for each test.