Configure JavaScript tests
Add the @web/test-runner and other related packages. We’re using
playwright as our runner because it bundles the necessary browsers and
this way we don’t have to worry about configuring the path to the
Chromium executable for the default test runner (which is a problem if
you install Chromium as a flatpak).

Also include sinon for spying on and mocking functions.