Add a component for generically handling data loading
Instead of implementing `src` attributes on all of our charts, we can
use this element (similar to a `<source>` element inside `<video>` or
`<picture>` elements) to load our data. We can implement handling of
`<chart-source>` elements in our base ChartElement class (or perhaps as
a mixin) and provide properties for data access, and then we don’t have
to implement any of this in our charts.