Insert styles and templates in the ChartElement base class
This should help reduce much of the code in chart subclasses because we
can just implement the getters for template and style and that will give
us the necessary shadow root contents. These are implemented as instance
properties (instead of static, the way Lit does it) so that we can
create mixins to provide common templates, like the D3 Margin
convention.