Rename the style property to CSS
HTMLElements already have a style property, so I can't use it for our
scoped style definition in subclasses. Rather than using `styles` the
way Lit does (which is too easily mistyped as `style`) I've changed it
to `css` and added a test to ensure that we don't insert a style
declaration when there are no styles for a component (which is how I
discovered the shadowed property).