Add tests for writing the parquet file during etl
Add some assertions to check that the parquet file is being created
during ETL. This is getting to be a bit much with the Zarr and database
stuff as well. If this works out, we should consider using just the
Parquet files for diag data and removing Zarr and Postgres from the
equation.