Separate each variable into separate Parquet partitioned file
Instead of combining all of the variables into a single parquet file for
the model, I've decided to separate them. Some of the variables, such as
wind, have columns that are unique to that variable; I think this is
reason enough to store them separately, otherwise all of the other
variables will just NaN values in those columns, which seems silly.

Some preliminary testing suggests this might actually save space.