Update diag.history to read from parquet
Update the tests and the implementation for retrieving historical data.
In the process, I simplified the response type. I think, in the future,
if we want to select unadjusted O - F or just the historical
observations, we should build that into the URL, rather than sending all
of the data back to the client and letting the client pick what to view.
This cuts down on data transfer and allows us to simplify our components
when they read data, since they don't have to parse a nested JavaScript
object.