Eliminate ".." from the parquet path
This looks like it fixes the missing bucket name issue with pyarrow
writing the Parquet files. Instead of building a relative path using a
Path object - which works fine on local filesystems - I use os.path to
strip the Zarr filename from the path and replace it with the Parquet
directory. This works locally and with S3.

I'm not entirely sure that I think this is a great solution, but adding
another path argument to diag.save() is a little cumbersome, and I
think, ultimately, we want to try eliminating the Zarr file and the
database in favor of just the Parquet files.