Point pytest to our tests directory (#357)
This speeds up test collection tremendously. Pytest takes about 2
minutes to check the entire project for tests, but if we point it at the
tests/ directory, it takes 0.11s to collect all the tests. Since we
don't have doctests in the code, there's no reason to have pytest check
the whole project.