Convert test_get_bounds to a class
Use the pattern for multiple assertions to test the return value from
test_get_bounds. I still find this a little hard to read (I wasn't sure
I remembered exactly what this function was doing, so I had to do a
little memory refresh). That probably means this needs a refactor so
that it's more clear what is happening.