Add a utility to help with S3 object renaming
I've created a script to do the bulk renaming that is hardcoded to our
bucket and timestamp pattern. Ideally, that wouldn't be the case but I
figure we can always make this script more flexible if the need arises.

I've also added a handy poetry entrypoint so that this can be run with
"poetry run".