Replace diag_dataset fixture with test_dataset in test_open_diagnostic
I'm trying to make what's happening in the tests a little more clear,
and also eliminate the diag_dataset fixture, which is not really
flexible enough for many tests. The test_dataset fixture is preferred
because it allows you to pass in data for the variables contained in a
diag dataset so that you can test more specific situations (such as
handling of the Analysis_Use_Flag column).