Convert test_load to a class without diag_file fixture
The TestLoad tests now follow the pattern for safely asserting multiple
conditions, so that one failed assertion doesn't mask other failures.
I've also converted it to use some module-level fixtures instead of the
diag_file fixture from conftest. test_extract is the only place we need
to generate the NetCDF files, so it makes more sense to keep this kind
of fixture out of conftest.