Convert the test_no_forecast_scalar test to a class
Separated the inputs into separate fixtures to make it a little easier
to identify and reuse the portions of test input data. Instead of
testing the entire dataset loaded from the test input, we just check the
forecast variables, because the purpose of the test is to ensure that
the forecast is calculated from the obs and obs-fcst variables when its
missing.