Remove diag_zarr_file fixture from test_diag
There was only one test outside of test_unified_graphics that was using
this fixture, and I think it makes more sense in this case to just rely
on the tmp_path fixture and generate the path string in the test setup.
Since so many of our tests are using the class pattern for multiple
asserts, a global function-scoped fixture with a path to the diag zarr
makes less sense. Especially since I want to convert the app fixture
(which relies on this fixture) to a module-scoped fixture for
test_unified_graphics so we don't have to keep recreating the app
object.