Remove S3 iterations of tests
Our code could use some clean-up so that we can better isolate testing
I/O for things like S3, but the real issue we're trying to fix is
eliminating the "../" token from the path being used as a key for the
Parquet files in S3. We don't strictly need S3 to test that that is the
case.