Explicitly provide the directory for Parquet files
Instead of trying to guess the right location for Parquet files based on
where we keep the Zarr file, we should just provide this explicitly. In
the event that a bad path breaks our ETL pipeline again, it will just be
a config change in production, not a code change. Additionally, this
relieves us of the burden of having to test the code that generates the
path and dealing with differences in allowed paths for file systems vs.
S3.