Use env variable to find Parquet files in routes
Instead of deriving the Parquet file path from the Zarr path, the route
reads in the FLASK_DIAG_PARQUET variable from the environment and passes
that explicitly to diag.history. diag.history uses that to find the
Parquet file for this model, and strips out `file://` if it's present
because Pandas is a bit silly about that.