Fix data type for is_used in test fixtures
When we process diag files in our pipeline, we convert the integers in
the diag files for is_used into booleans, so should treat them the same
way in our test fixtures. I think this is a sign that our test fixtures
are poorly set up, since they can get out of sync with reality. The
result of this problem was that either the application worked, or our
tests passed. Prior to this change, we needed to compare is_used to a
boolean for Parquet files generated with our ETL code, but to an integer
for test files generated with our fixtures.